Readme/newcomer infos (#354)
* Added informations about typescript usage

I admit there's not much there, but as a newcomer I spent some time looking around for it. 
Turns out it's so simple it works by default :D But you have to know it.

* Surfacing the input/output file default config

As a newcomer, I missed this information as it was a bit buried in a big paragraph. I think this might make it stand out a bit more.

* Infos about default input/output

I think it has Its place in the installation phase as it's required for everything to work and has to be set up anyway.

* Cli options at the end of usage chapter

It seems more natural to me to end this way, but really that's subjective.

* Tidy up

* Added typescript to features

It's a very nice feature, it deserves Its place there (I think)

* Added "types" to specifying builds chapter

* Apply suggestions from code review

Co-Authored-By: isthatcentered <e.peninb@gmail.com>

* Apply suggestions from code review

Co-Authored-By: isthatcentered <e.peninb@gmail.com>

* Added latest full cli --help output